refactor: 카테고리 폼 분리
